requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
click==8.1.3; python_version >= '3.7'
colorama==0.4.6; platform_system == 'Windows'
flask-mysqldb==1.0.1
flask==2.2.3
importlib-metadata==6.1.0; python_version < '3.10'
itsdangerous==2.1.2; python_version >= '3.7'
jinja2==3.1.2; python_version >= '3.7'
markupsafe==2.1.2; python_version >= '3.7'
mysqlclient==2.1.1; python_version >= '3.5'
python-dotenv==1.0.0
werkzeug==2.2.3; python_version >= '3.7'
zipp==3.15.0; python_version >= '3.7'
